[clean]検索ページの処理をカスタムフックに切り出し
